Added Scene Management Script and made changes on Exit Script
